warroom/requirements.txt
pipecat-ai[websocket,deepgram,cartesia,silero]==0.0.108
python-dotenv>=1.0.0
